fix: fixed status bar color hiding top app bar
