+++ app/src/main/java/quo/yandex/financialawareness/MainActivity.kt
@@ -5,45 +5,33 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.asPaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.safeDrawingPadding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.statusBarsPadding
-import androidx.compose.material3.CenterAlignedTopAppBar
-import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.foundation.layout.statusBars
 import androidx.compose.material3.NavigationBar
 import androidx.compose.material3.NavigationBarDefaults
 import androidx.compose.material3.NavigationBarItem
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarColors
-import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.core.view.WindowCompat
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import quo.yandex.financialawareness.presentation.screens.AccountScreen
-import quo.yandex.financialawareness.presentation.screens.CategoriesScreen
-import quo.yandex.financialawareness.presentation.screens.ExpensesScreen
-import quo.yandex.financialawareness.presentation.screens.IncomeScreen
-import quo.yandex.financialawareness.presentation.screens.SettingsScreen
+import quo.yandex.financialawareness.presentation.navigation.AppNavHost
+import quo.yandex.financialawareness.presentation.navigation.Destination
 import quo.yandex.financialawareness.presentation.ui.theme.FinancialAwarenessTheme
 
 class MainActivity : ComponentActivity() {
@@ -54,19 +42,6 @@ class MainActivity : ComponentActivity() {
             AppScreen()
         }
     }
-}
-
-enum class Destination(
-    val route: String,
-    val label: String,
-    val icon: Int,
-    val contentDescription: String
-) {
-    EXPENSES("expanses", "Расходы", R.drawable.downtrend, "Расходы"),
-    INCOME("income", "Доходы", R.drawable.uptrend, "Доходы"),
-    ACCOUNT("account", "Счёт", R.drawable.calculator, "Счёт"),
-    CATEGORIES("categories", "Статьи", R.drawable.bar_chart_side, "Статьи"),
-    SETTINGS("settings", "Настройки", R.drawable.settings, "Настройки"),
 }
 
 @Preview
@@ -101,6 +76,7 @@ fun AppScreen(modifier: Modifier = Modifier) {
                 }
             },
         ) { innerPadding ->
+            TransparentStatusBarBackground(Color(0xFF2AE881))
             Box(modifier = Modifier.padding(innerPadding)) {
                 AppNavHost(navController, startDestination)
             }
@@ -109,25 +85,14 @@ fun AppScreen(modifier: Modifier = Modifier) {
 }
 
 @Composable
-fun AppNavHost(
-    navController: NavHostController,
-    startDestination: Destination,
-    modifier: Modifier = Modifier,
-) {
-    NavHost(
-        navController,
-        startDestination = startDestination.route
-    ) {
-        Destination.entries.forEach { destination ->
-            composable(destination.route) {
-                when (destination) {
-                    Destination.EXPENSES -> ExpensesScreen(modifier)
-                    Destination.INCOME -> IncomeScreen(modifier)
-                    Destination.ACCOUNT -> AccountScreen(modifier)
-                    Destination.CATEGORIES -> CategoriesScreen(modifier)
-                    Destination.SETTINGS -> SettingsScreen(modifier)
-                }
-            }
-        }
-    }
+fun TransparentStatusBarBackground(color: Color) {
+    val topPadding = WindowInsets.statusBars
+        .asPaddingValues()
+        .calculateTopPadding()
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(topPadding)
+            .background(color)
+    )
 }
